+++ packages/vchart-types/types/plugin/components/tooltip-handler/interface/index.d.ts
@@ -1,2 +1,15 @@
-export * from './style';
-export * from './common';
+import type { TooltipAttributes } from '@visactor/vrender-components';
+import type { ITextAttribute, IFillStyle } from '@visactor/vrender-core';
+export interface ITooltipTextStyle extends Partial<ITextAttribute & IFillStyle> {
+    spacing?: number;
+    multiLine?: boolean;
+    maxWidth?: number;
+    autoWidth?: boolean;
+}
+export interface ITooltipAttributes extends TooltipAttributes {
+    align?: 'left' | 'right';
+}
+export interface ITooltipHandlerOptions {
+    offsetX?: number;
+    offsetY?: number;
+}