+++ packages/vchart/src/chart/pictogram/interface.ts
@@ -1,0 +1,7 @@
+import { IPictogramSeriesSpec } from '../../series/pictogram/interface';
+import type { IChartExtendsSeriesSpec, IChartSpec } from '../../typings/spec/common';
+
+export interface IPictogramChartSpec extends IChartSpec, IChartExtendsSeriesSpec<any> {
+  type: 'pictogram';
+  series?: IPictogramSeriesSpec[];
+}