+++ packages/vchart/src/plugin/components/tooltip-handler/utils/attribute.ts
@@ -3,22 +3,70 @@ import type {
   TooltipRowAttrs,
   TooltipRowStyleAttrs,
   TooltipSymbolAttrs,
-  TooltipTextAttrs
+  TooltipTextAttrs,
+  TooltipRichTextAttrs
 } from '@visactor/vrender-components';
-import type { IPadding, ITooltipActual } from '../../../../typings';
+import type { IPadding, ITooltipActual, MaybeArray } from '../../../../typings';
 import type { ITooltipAttributes, ITooltipTextStyle } from '../interface';
 import { isValid, maxInArray, normalizePadding } from '@visactor/vutils';
 import { mergeSpec } from '@visactor/vutils-extension';
 import { normalizeLayoutPaddingSpec } from '../../../../util/space';
-import { measureTooltipText } from './common';
 import type { ITheme } from '../../../../theme';
 import type { ITooltipSpec, ITooltipTextTheme, ITooltipTheme } from '../../../../component/tooltip';
 import { token } from '../../../../theme/token';
+// eslint-disable-next-line no-duplicate-imports
+import { getRichTextBounds } from '@visactor/vrender-core';
+// eslint-disable-next-line no-duplicate-imports
+import type { IRichTextCharacter, IRichTextParagraphCharacter } from '@visactor/vrender-core';
 
 const DEFAULT_TEXT_ATTRIBUTES: Partial<ITooltipTextStyle> = {
   fontFamily: token.fontFamily,
   spacing: 10,
   wordBreak: 'break-word'
+};
+
+interface ITooltipTextInfo {
+  width: number;
+  height: number;
+  text: MaybeArray<number> | MaybeArray<string> | TooltipRichTextAttrs;
+}
+
+/** 测量 tooltip 标签文本 */
+export const measureTooltipText = (text: string | TooltipRichTextAttrs, style: ITooltipTextStyle): ITooltipTextInfo => {
+  let textLines: string[] | TooltipRichTextAttrs;
+  let textConfig: IRichTextCharacter[];
+  if (!((text as TooltipRichTextAttrs)?.type === 'rich' || (text as TooltipRichTextAttrs)?.type === 'html')) {
+    text = (text ?? '').toString();
+    if (style.multiLine) {
+      textLines = text.split('\n');
+      textLines = textLines.map((line, i) => (i < (textLines as string[]).length - 1 ? line + '\n' : line));
+    } else {
+      textLines = [text];
+    }
+    textConfig = textLines.map(
+      line =>
+        ({
+          ...style,
+          text: line
+        } as unknown as IRichTextParagraphCharacter)
+    );
+  } else {
+    textConfig = (text as TooltipRichTextAttrs).text as IRichTextCharacter[];
+    textLines = text as TooltipRichTextAttrs;
+  }
+
+  const bound = getRichTextBounds({
+    wordBreak: (style as any).wordBreak ?? 'break-word',
+    maxWidth: style.maxWidth ? style.maxWidth : undefined,
+    width: 0,
+    height: 0,
+    textConfig: textConfig
+  });
+  return {
+    width: bound.width(),
+    height: bound.height(),
+    text: textLines
+  };
 };
 
 export function getTextAttributes(
@@ -76,16 +124,7 @@ export const getTooltipAttributes = (
   globalTheme: ITheme
 ): ITooltipAttributes => {
   const { style = {}, enterable, transitionDuration } = spec;
-  const {
-    panel = {},
-    titleLabel,
-    shape,
-    keyLabel,
-    valueLabel,
-    spaceRow: commonSpaceRow,
-    maxContentHeight,
-    align
-  } = style;
+  const { panel = {}, titleLabel, shape, keyLabel, valueLabel, spaceRow: commonSpaceRow, align } = style;
   const padding = normalizePadding(panel.padding);
   const paddingSpec = normalizeLayoutPaddingSpec(panel.padding) as IPadding;
 
@@ -131,8 +170,6 @@ export const getTooltipAttributes = (
     keyWidth: 0,
     valueWidth: 0,
 
-    maxContentHeight,
-
     enterable,
     transitionDuration,
     align
@@ -142,8 +179,6 @@ export const getTooltipAttributes = (
 
   let panelWidth = paddingSpec.left + paddingSpec.right;
   let panelHeight = paddingSpec.top + paddingSpec.bottom;
-  /** dom tooltip 的高度。由于 canvas tooltip 不支持滚动条，dom tooltip 单独计算高度 */
-  let panelDomHeight = paddingSpec.top + paddingSpec.bottom;
 
   // calculate content
   let contentMaxWidth = 0;
@@ -245,7 +280,6 @@ export const getTooltipAttributes = (
       return itemAttrs;
     });
     panelHeight += contentHeight;
-    panelDomHeight += Math.min(contentHeight, maxContentHeight ?? Infinity);
 
     maxKeyWidth = keyWidths.length ? maxInArray(keyWidths) : 0; // name 需要对齐
     maxAdaptiveKeyWidth = adaptiveKeyWidths.length ? maxInArray(adaptiveKeyWidths) : 0;
@@ -298,7 +332,6 @@ export const getTooltipAttributes = (
     titleHeightWithSpace = titleMaxHeight + (hasContent ? attributes.title.spaceRow : 0);
   }
   panelHeight += titleHeightWithSpace;
-  panelDomHeight += titleHeightWithSpace;
   attributes.title.width = titleMaxWidth;
   attributes.title.height = titleMaxHeight;
 
@@ -333,6 +366,5 @@ export const getTooltipAttributes = (
 
   attributes.panel.width = panelWidth;
   attributes.panel.height = panelHeight;
-  attributes.panelDomHeight = panelDomHeight;
   return attributes;
 };